requirements.txt
certifi==2018.4.16
future==0.16.0
numpy==1.14.5
opencv-python==3.4.1.15
Pillow==5.2.0
python-telegram-bot==10.1.0
